backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
greenlet>=3.0.0
pymongo>=4.6.0
motor>=3.3.2
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

